requirements.txt
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
audioread==3.0.1
Brlapi==0.8.5
Brotli==1.1.0
catfish==4.16.4
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.0
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
debtags==2.1
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
Django==5.0.3
django-rest-framework==0.1.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
executing==2.0.1
fonttools==4.50.0
html5lib==1.1
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kali-tweaks==2023.3.2
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
lightdm-gtk-greeter-settings==1.2.3
llvmlite==0.42.0
louis==3.29.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
mechanize==0.4.9
msgpack==1.0.8
mutagen==1.46.0
nest-asyncio==1.6.0
netifaces==0.11.0
numba==0.59.1
numpy==1.26.4
onboard==1.4.1
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.25.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.47.0.dev0
pyparsing==3.1.2
pysmbc==1.0.25.1
python-apt==2.7.6
python-dateutil==2.9.0.post0
python-debian==0.1.49
pytz==2024.1
pyxdg==0.28
pyzmq==25.1.2
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.4.4
stack-data==0.6.3
threadpoolctl==3.4.0
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.4
xdg==5
yt-dlp==2024.3.10
